requirements.txt
numpy==1.26.4
fastapi~=0.115.9
uvicorn
transformers~=4.51.1
torch
pandas~=2.2.3
sentence-transformers~=4.0.2
faiss-cpu
scikit-learn~=1.6.1
jupyterlab
requests~=2.32.3
chromadb~=1.0.4
python-dotenv~=1.1.0
pydantic~=2.11.3
SQLAlchemy~=2.0.40